+++ src/core.ts
@@ -1,2 +1,2 @@
-export { HlsDownloader, type GetAdditionalOption } from '@hls-downloader/core';
+export { HlsDownloader, type GetAdditionalOptions } from '@hls-downloader/core';
 export { default } from '@hls-downloader/core';